[Refactor] finish refactoring
