+++ scene_objects/mesh.cpp
