+++ scene_objects/sphere.cpp
