+++ scene_objects/triangle.h
@@ -1,9 +1,9 @@
 #ifndef TRIANGLE_H
 #define TRIANGLE_H
 
-#include "vec3.h"
-#include "ray.h"
-#include "object.h"
+#include "tools/vec3.h"
+#include "tools/ray.h"
+#include "tools/object.h"
 
 class Triangle : public Object{
     public:
