+++ tools/ray.cpp
