+++ tools/ray.h
