+++ tools/vec3.cpp
